Revert "Add functions"
