--- code/TODO
@@ -1,48 +1,0 @@
-# Code reorganization
-
-Create modules for reading FASTA and GFA files;
-Check validity of GFA file
-
-# Parameters
-
-Read parameters for tuning and preprocessing/running from YAML files
-
-For plasbin_flows, read input from a CSV file
-
-# Code cleaning
-
-Clean up the code in modules
-- ground_truth
-- seeds
-- gc_content
-- get_data
-- model_setup
-Uniformize comments
-Uniformize indentation
-Create functions
-Break large blocks of code
-Move hard-doced constants into parameters
-
-Create functions in plasbin_flow
-
-# Code extension
-
-Module gc_content:
-- create the GC intervals file
-- check that an input GC content intervals file is correct
-- read GC probabilities and GC intervals files
-
-Module ground_truth:
-- check validity of a ground truth file
-- read ground truth file
-
-# Logging and error handling
-
-Improve logging
-- all messages should be redirected to log file
-- error messages should be printed to stderr
-- errors should stop scripts
-
-# Rewrite README.md
-- Add list of modules needed and versions
-- Describe new commands using only plasbin_utils and plasbin_flow
